first steps for the repository only approach
